+++ tests/includes/collection/class-test-followers.php
@@ -359,23 +359,6 @@ class Test_Followers extends \WP_UnitTestCase {
 	}
 
 	/**
-	 * Tests scheduling of migration.
-	 *
-	 * @covers ::maybe_migrate
-	 */
-	public function test_migration_scheduling() {
-		update_option( 'activitypub_db_version', '0.0.1' );
-
-		\Activitypub\Migration::maybe_migrate();
-
-		$schedule = \wp_next_scheduled( 'activitypub_migrate', array( '0.0.1' ) );
-		$this->assertNotFalse( $schedule );
-
-		// Clean up.
-		delete_option( 'activitypub_db_version' );
-	}
-
-	/**
 	 * Data provider for migration test scenarios.
 	 *
 	 * @return array[]